fix:onboarding screens displaying on first launch of the app in a phone
