--- redux/reducers/index.ts
@@ -1,10 +1,0 @@
-import { combineReducers } from "@reduxjs/toolkit";
-
-import onboardingReducer from "./obReducer";
-
-const rootReducer = combineReducers({
-
-    onboarding: onboardingReducer,
-});
-
-export default rootReducer;
